requirements.txt
python-gitlab==1.1.0
